requirements.txt
ultralytics==8.3.23
opencv-python==4.10.0.84
numpy==1.26.4
tqdm==4.66.4
streamlit==1.36.0
moviepy==1.0.3
Pillow==10.4.0